--- src/config/posts/initialData.ts
@@ -1,7 +1,0 @@
-import { NewPost } from '../../entities/posts/model/NewPost';
-
-export const initialNewPost: NewPost = {
-  title: '',
-  body: '',
-  userId: 1,
-} as const;
