+++ src/features/dialog/ui/BaseDialog.tsx
