+++ src/features/posts/api/index.ts
@@ -1,0 +1,2 @@
+export { useAddPost, useDeletePost, useUpdatePost } from './usePostsMutations';
+export { useQueryPostById, useQueryPosts } from './usePostsQueries';
