+++ src/features/posts/config/initialData.ts
@@ -1,0 +1,7 @@
+import { NewPost } from '../../../entities/posts/model/NewPost';
+
+export const initialNewPost: NewPost = {
+  title: '',
+  body: '',
+  userId: 1,
+} as const;
