+++ src/features/posts/model/index.ts
@@ -1,0 +1,5 @@
+export { usePostsStore, usePostsStoreSelector } from './usePostsStore';
+export {
+  useSelectedPostStore,
+  useSelectedPostStoreSelector,
+} from './useSelectedPostStore';
