+++ src/features/users/api/index.ts
@@ -1,0 +1,2 @@
+export { useSelectedUserMutation } from './useUsersMutations';
+export { useQueryUser, useQueryUsers } from './useUsersQueries';
