--- src/stores/comments/useNewCommentStore.ts
@@ -1,17 +1,0 @@
-import type { Comment } from '../../types';
-import { createStoreSelector } from '../../shared/lib';
-import { create } from 'zustand';
-import { initialComment } from '../../config/comments/initialData';
-
-export const useNewCommentStore = create<{
-  newComment: Partial<Comment> | null;
-  updateNewComment: (comment: Partial<Comment>) => void;
-  resetNewComment: () => void;
-}>((set) => ({
-  newComment: initialComment,
-  updateNewComment: (comment) => set({ newComment: comment }),
-  resetNewComment: () => set({ newComment: initialComment }),
-}));
-
-export const useNewCommentStoreSelector =
-  createStoreSelector(useNewCommentStore);
