--- src/stores/comments/useSelectedCommentStore.ts
@@ -1,15 +1,0 @@
-import type { Comment } from '../../types';
-import { create } from 'zustand';
-import { createStoreSelector } from '../../shared/lib';
-
-export const useSelectedCommentStore = create<{
-  selectedComment: Comment | null;
-  setSelectedComment: (comment: Comment) => void;
-}>((set) => ({
-  selectedComment: null,
-  setSelectedComment: (comment) => set({ selectedComment: comment }),
-}));
-
-export const useSelectedCommentStoreSelector = createStoreSelector(
-  useSelectedCommentStore,
-);
